move to map2tiles
